chore(pattern): move `RandomPatternTest` to `patterns` directory
